Make plugin data regex as flexible as WP core
The updater would only allow zero or more spaces before Version when trying to get the remote version. This PR changes the regex to match the one used on WP get_file_data(), which allows zero or more spaces, tabs, asterisks, hashes and at-symbols. This means that plugin file data contained within a DocBlock (therefore starting with a space, an asterisk, then another space before Version), such as found in the WordPress Plugin Boilerplate project, are recognised by the updater.