Use $wp_filesystem to rename the plugin folder.
Instead using native file and folder manipulation functions, here, rename(), we should be using the WP_Filesystem. This has far better support across different systems and setups.

By the time we get to renaming the source folder, the global $wp_filesystem has already been setup, so we can simply re-use that to effectively rename the folder by moving it instead. Behind the scenes, WP_Filesystem will try to use @rename(), and if that doesn't work, it will copy the folder to a new location and then delete the original.