Needs PHP Zip extension.
WordPress Version: 4.6.1
PHP Version:       7.0.11-1+deb.sury.org~xenial+1

on Ubuntu 16.04 and I had the plugin failing with `The package could not be installed. PCLZIP_ERR_BAD_FORMAT (-10) : Invalid archive structure` until install `php-zip`. Maybe there could be a better error message because is was just a guess and the message itself does not really point to the solution.